LPS-166923 Drop the Many-to-Many table only when it has been created

+++ modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectRelationshipLocalServiceImpl.java
@@ -279,7 +279,9 @@ public class ObjectRelationshipLocalServiceImpl
 					objectRelationship.getType(),
 					ObjectRelationshipConstants.TYPE_MANY_TO_MANY)) {
 
-			runSQL("drop table " + objectRelationship.getDBTableName());
+			if (Validator.isNotNull(objectRelationship.getDBTableName())) {
+				runSQL("drop table " + objectRelationship.getDBTableName());
+			}
 
 			ObjectRelationship reverseObjectRelationship =
 				fetchReverseObjectRelationship(objectRelationship, true);
